sample code refactor
