+++ sample/src/main/java/com/sampleapp/functions/network/DemoNetworkFragment.kt
@@ -8,7 +8,9 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.sampleapp.R
 import com.sampleapp.databinding.FragmentDemoNetworkBinding
-import com.sampleapp.functions.network.internal.NetworkViewModel
+import com.sampleapp.functions.network.internal.custom.CustomViewModel
+import com.sampleapp.functions.network.internal.ktor.KtorViewModel
+import com.sampleapp.functions.network.internal.okhttp.OkhttpViewModel
 
 class DemoNetworkFragment : Fragment(R.layout.fragment_demo_network) {
 
@@ -16,7 +18,9 @@ class DemoNetworkFragment : Fragment(R.layout.fragment_demo_network) {
     private val binding
         get() = _binding!!
 
-    private val networkViewModel: NetworkViewModel by viewModels()
+    private val okhttpViewModel: OkhttpViewModel by viewModels()
+    private val ktorViewModel: KtorViewModel by viewModels()
+    private val customViewModel: CustomViewModel by viewModels()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         _binding = FragmentDemoNetworkBinding.inflate(inflater, container, false)
@@ -30,12 +34,12 @@ class DemoNetworkFragment : Fragment(R.layout.fragment_demo_network) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        binding.postCall.setOnClickListener { networkViewModel.post() }
-        binding.getCall.setOnClickListener { networkViewModel.get() }
-        binding.getCallKtor.setOnClickListener { networkViewModel.getKtor() }
-        binding.postCallKtor.setOnClickListener { networkViewModel.postKtor() }
-        binding.xmlCall.setOnClickListener { networkViewModel.xml() }
-        binding.formEncodedCall.setOnClickListener { networkViewModel.form() }
-        binding.customTrace.setOnClickListener { networkViewModel.customTrace() }
+        binding.postCall.setOnClickListener { okhttpViewModel.post() }
+        binding.getCall.setOnClickListener { okhttpViewModel.get() }
+        binding.getCallKtor.setOnClickListener { ktorViewModel.get() }
+        binding.postCallKtor.setOnClickListener { ktorViewModel.post() }
+        binding.xmlCall.setOnClickListener { okhttpViewModel.xml() }
+        binding.formEncodedCall.setOnClickListener { okhttpViewModel.form() }
+        binding.customTrace.setOnClickListener { customViewModel.customTrace() }
     }
 }
