+++ sample/src/main/java/com/sampleapp/functions/network/internal/ktor/PostRequestBody.kt
@@ -1,0 +1,12 @@
+package com.sampleapp.functions.network.internal.ktor
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class PostRequestBody(
+    @SerialName("user")
+    val user: String,
+    @SerialName("email")
+    val email: String,
+)
