+++ sample/src/main/java/com/sampleapp/functions/network/internal/okhttp/ApiService.kt
@@ -1,4 +1,4 @@
-package com.sampleapp.functions.network.internal.core
+package com.sampleapp.functions.network.internal.okhttp
 
 import okhttp3.RequestBody
 import retrofit2.http.Body
