+++ sample/src/main/java/com/sampleapp/functions/network/internal/okhttp/NetworkCalls.kt
@@ -1,4 +1,4 @@
-package com.sampleapp.functions.network.internal.core
+package com.sampleapp.functions.network.internal.okhttp
 
 import android.util.Log
 import com.squareup.moshi.JsonAdapter
